Добавляет правки
